Mail Summarization
Added API to queue and summarize mails into one mail.
Mails can be summarized by category or by the
combination of a category and an entity.
Users can configure which of those strategies to choose.
The caller of the Mail Plugin API has to decide, which mails
should be queued and which mails should be sent out directly.

Also updated the simple-java-mail dependency to be compatible
with SCM 3.x

Co-authored-by: René Pfeuffer<rene.pfeuffer@cloudogu.com>